Move fundingOfferHistory sync-schema to separate file
