+++ workers/loc.api/sync/schema/sync-schema/funding-offer-history.js
@@ -1,0 +1,20 @@
+'use strict'
+
+const TABLES_NAMES = require('../tables-names')
+const ALLOWED_COLLS = require('../allowed.colls')
+const COLLS_TYPES = require('../colls.types')
+
+const { getModelOf } = require('../models')
+
+module.exports = {
+  name: ALLOWED_COLLS.FUNDING_OFFER_HISTORY,
+  maxLimit: 10000,
+  dateFieldName: 'mtsUpdate',
+  symbolFieldName: 'symbol',
+  sort: [['mtsUpdate', -1]],
+  hasNewData: false,
+  start: [],
+  isSyncRequiredAtLeastOnce: false,
+  type: COLLS_TYPES.INSERTABLE_ARRAY_OBJECTS,
+  model: getModelOf(TABLES_NAMES.FUNDING_OFFER_HISTORY)
+}
